pyproject.toml
[tool.poetry]
name = "retrain_bert"
version = "0.1.0"
description = ""
authors = ["Julius Juodagalvis <julius.juodagalvis@gmail.com>"]
packages = [
    { include = "retrain_bert", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^2.2.2"
transformers = "^4.41.0"
scikit-learn = "^1.5.0"
seaborn = "^0.13.0"
pyarrow = "^14.0.2"
azure-identity = "^1.16.0"
azureml-mlflow = "^1.56.0"
mlflow = "^2.11.0"
azureml = "^0.2.7"
azure-ai-ml = "^1.16.0"
tqdm = "^4.66.4"
pyodbc = "^5.1.0"
sqlalchemy = "^2.0.31"
torch = "^2.3.1"
lightning = "^2.3.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
ipython = "^8.24.0"
ipykernel = "^6.29.4"
black = "^23.12.1"
iprogress = "^0.4"
ipywidgets = "^8.1.2"
poetry-plugin-up = "^0.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
